+++ packages/local-prover/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- Remove `toJSON` and all `fromJSON` methods in favor of custom `bigint-json-serialization`
 - Add local prover for subtree update
 - Refactor joinsplit prover to take circuit files as class fields not function args
 - Rename all "flax" instances to "nocturne"